Restructuring and fixing of garbled output
I restructured the code of the 15bit value logic into their own
little data type to make it harder to make mistakes.

In the process of doing so I found out I misused the Out operation since
it can also refer to registers. Instead of reading them I just treated
the register address as a char; hence the garbled output.