Got the teleporter challenge working
It took a stupid amount of time but I figured out how to circumvent the
teleporter. Luckily r7 is not used in any other code except for the
self check which makes things a bit easier.
You have to
- use the teleporter with a r7 value != 0 (manipulate the register).
- see to which instructions the code jumps to validate.
- analyse these instructions.
- reverse engineer the check function.
- run the check function with different r7 values until you find the
  correct one

When you found the correct value for r7 to use you have to
- play the game again.
- set the r7 value to the computed value you got.
- patch the words in memory corresponding to the check of the r7 value
  with NOOP words (21)
- use the teleporter

Pretty brutal puzzle. No idea which mad man came up with this.

Im now pretty comfortable at reading SynacorVM assembly and
learned a lot about debugging rust code!