Completed the synacor challenge
The last problem was a Math-Maze which I wrote a program for that is not
pretty and definitely could use some rework.

After finding the unique shortest path through the Math-Maze you'll find
a mirror and the last keyword. Since you look through the mirror at the
keyword you have to "mirror" it before putting it into the synacore
homepage!

Cleaned up the VM code a bit so you know have to type in the secret
passphrase to fix the teleporter code.

The play_game.sh script is now able to play the game through in one fast
swuuuuush.