Cleaning up work
Some smaller cleanups like putting the u15 data type in it's own little
library.
Renaming some things and a little bit of cleanup in the VM, the
coinsolver, the r7 calculator.

The mazesolver still needs to be overhauled after I get a better grasp
at boxed closures and how to pass them around and use them.